--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@ dist/
 # For "cabal sandbox"
 .cabal-sandbox
 cabal.sandbox.config
+dist-newstyle