--- a/proto-lens-benchmarks/benchmarks/large_fields.proto
+++ b/proto-lens-benchmarks/benchmarks/large_fields.proto
@@ -1,0 +1,7 @@
+syntax = "proto3";
+
+package large_fields;
+
+message Foo {
+  string str_field = 1;
+}