--- a/proto-lens-benchmarks/proto-lens-benchmarks.cabal
+++ b/proto-lens-benchmarks/proto-lens-benchmarks.cabal
@@ -1,4 +1,4 @@
-cabal-version: 1.24
+cabal-version: 2.0
 
 -- This file has been generated from package.yaml by hpack version 0.34.4.
 --
@@ -16,6 +16,7 @@ license:        BSD3
 license-file:   LICENSE
 build-type:     Custom
 extra-source-files:
+    benchmarks/large_fields.proto
     benchmarks/nested.proto
     benchmarks/packing.proto
     benchmarks/unused_fields.proto
@@ -47,10 +48,37 @@ library
     , text
   default-language: Haskell2010
 
+benchmark large-fields
+  type: exitcode-stdio-1.0
+  main-is: LargeFields.hs
+  other-modules:
+      Proto.LargeFields
+      Proto.LargeFields_Fields
+  autogen-modules:
+      Proto.LargeFields
+      Proto.LargeFields_Fields
+  hs-source-dirs:
+      benchmarks
+  ghc-options: -O2 -rtsopts
+  build-depends:
+      base
+    , criterion
+    , deepseq
+    , lens-family
+    , lens-family-core
+    , proto-lens
+    , proto-lens-benchmarks
+    , proto-lens-runtime
+    , text
+  default-language: Haskell2010
+
 benchmark nested
   type: exitcode-stdio-1.0
   main-is: Nested.hs
   other-modules:
+      Proto.Nested
+      Proto.Nested_Fields
+  autogen-modules:
       Proto.Nested
       Proto.Nested_Fields
   hs-source-dirs:
@@ -74,6 +102,9 @@ benchmark packing
   other-modules:
       Proto.Packing
       Proto.Packing_Fields
+  autogen-modules:
+      Proto.Packing
+      Proto.Packing_Fields
   hs-source-dirs:
       benchmarks
   ghc-options: -O2 -rtsopts
@@ -93,6 +124,9 @@ benchmark unused-fields
   type: exitcode-stdio-1.0
   main-is: UnusedFields.hs
   other-modules:
+      Proto.UnusedFields
+      Proto.UnusedFields_Fields
+  autogen-modules:
       Proto.UnusedFields
       Proto.UnusedFields_Fields
   hs-source-dirs: