@blackgnezdo 2022-08-29
I measured this on my end too (OpenBSD, ghc-9.4.2, AMD Ryzen 7 5700G)

Before
```
benchmarking 10M(10MB)/decode/whnf
time                 5.495 ms   (5.388 ms .. 5.598 ms)
                     0.996 R²   (0.993 R² .. 0.997 R²)
mean                 5.571 ms   (5.461 ms .. 5.726 ms)
std dev              394.8 μs   (280.5 μs .. 611.7 μs)
variance introduced by outliers: 42% (moderately inflated)

benchmarking 10M(10MB)/decode/nf
time                 5.497 ms   (5.378 ms .. 5.607 ms)
                     0.996 R²   (0.993 R² .. 0.997 R²)
mean                 5.562 ms   (5.451 ms .. 5.733 ms)
std dev              416.6 μs   (269.4 μs .. 652.5 μs)
variance introduced by outliers: 45% (moderately inflated)

Benchmark large-fields: FINISH
```

After

```
benchmarking 10M(10MB)/decode/whnf
time                 3.685 ms   (3.544 ms .. 3.803 ms)
                     0.977 R²   (0.960 R² .. 0.987 R²)
mean                 4.170 ms   (3.930 ms .. 4.483 ms)
std dev              874.3 μs   (638.7 μs .. 1.244 ms)
variance introduced by outliers: 88% (severely inflated)

benchmarking 10M(10MB)/decode/nf
time                 3.744 ms   (3.608 ms .. 3.871 ms)
                     0.977 R²   (0.960 R² .. 0.987 R²)
mean                 4.050 ms   (3.850 ms .. 4.393 ms)
std dev              843.3 μs   (567.7 μs .. 1.257 ms)
variance introduced by outliers: 88% (severely inflated)

Benchmark large-fields: FINISH
```