@4eUeP 2022-08-30
@blackgnezdo 

> One quick request: could you rearrange the commits so the benchmark is before the fix? This makes it easy to have a benchmark comparison run should somebody want to check your and my runs.

Sounds reasonable. DONE